Update temperature to float

+++ llama-index-packs/llama-index-packs-mixture-of-agents/llama_index/packs/mixture_of_agents/base.py
@@ -49,7 +49,7 @@ class MixtureOfAgentWorkflow(Workflow):
         reference_llms: List[LLM],
         num_layers: int = 3,
         max_tokens: int = 2048,
-        temperature: int = 0.7,
+        temperature: float = 0.7,
         **kwargs,
     ):
         super().__init__(**kwargs)
